+++ NEWS.md
@@ -1,3 +1,11 @@
+# softwareRisk 0.2.0
+
+- Added support to calculate a power mean version of the function's risk score.
+
+- Added the function "risk_form_variance_share_fun" to evaluate how much uncertainty
+in each node's risk score is due to the selection of the functional form to calculate
+the risk score ("additive" vs "power mean").
+
 # softwareRisk 0.1.0
 
 - Initial CRAN release.